feat: 플레이어의 수와 플레이어 이름 중복 여부 검증 로직을 Players 클래스로 이동
