+++ src/main/kotlin/blackjack/model/Players.kt
@@ -2,7 +2,8 @@ package blackjack.model
 
 class Players(val value: List<Player>) {
     init {
-        require(value.isNotEmpty()) { }
+        require(value.map { it.gameInfo.name }.distinct().size == value.size) { EXCEPTION_DUPLICATED_PLAYERS }
+        require(value.size in PLAYER_NUM_RANGE) { EXCEPTION_NUMBER_OF_PLAYERS.format(value.size) }
         initializeCards()
     }
 
@@ -13,6 +14,14 @@ class Players(val value: List<Player>) {
     }
 
     companion object {
+        private const val MINIMUM_NUMBER_OF_PLAYERS = 2
+        private const val MAXIMUM_NUMBER_OF_PLAYERS = 8
+        private val PLAYER_NUM_RANGE = MINIMUM_NUMBER_OF_PLAYERS..MAXIMUM_NUMBER_OF_PLAYERS
+        private const val EXCEPTION_NUMBER_OF_PLAYERS =
+            "플레이어의 수로 %d를 입력했습니다. 플레이어 수는 ${MINIMUM_NUMBER_OF_PLAYERS}부터 ${MAXIMUM_NUMBER_OF_PLAYERS}까지 가능합니다."
+        private const val EXCEPTION_DUPLICATED_PLAYERS =
+            "중복된 플레이어의 이름을 입력하셨습니다."
+
         fun of(
             playerNames: List<String>,
             onInputDecision: (String) -> String,
