+++ src/test/kotlin/blackjack/model/DealerTest.kt
@@ -3,39 +3,15 @@ package blackjack.model
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
-import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.MethodSource
-import org.junit.jupiter.params.provider.ValueSource
 import java.util.stream.Stream
 
 class DealerTest {
-    @ParameterizedTest
-    @ValueSource(ints = [1, 9])
-    fun `참여자 수가 허용 범위를 벗어나면, 예외를 발생시킨다`(numberOfPlayers: Int) {
-        val players = createMultiplePlayers(numberOfPlayers)
-        val exception =
-            assertThrows<IllegalArgumentException> {
-                Dealer(players)
-            }
-        assertThat(exception.message).isEqualTo("플레이어의 수로 ${players.size}를 입력했습니다. 플레이어 수는 2부터 8까지 가능합니다.")
-    }
-
-    @Test
-    fun `참여자 이름이 중복이 있으면, 예외를 발생시킨다`() {
-        val players = listOf("케이엠", "케이엠", "해음")
-        val exception =
-            assertThrows<IllegalArgumentException> {
-                Dealer(players)
-            }
-        assertThat(exception.message).isEqualTo("중복된 플레이어의 이름을 입력하셨습니다.")
-    }
-
     @Test
     fun `딜러 카드의 총합을 계산한다`() {
         val dealer =
             Dealer(
-                players = createMultiplePlayers(2),
                 gameInfo = createHitDealerGameInfo(),
             )
         assertThat(dealer.gameInfo.total).isEqualTo(16)
@@ -45,7 +21,6 @@ class DealerTest {
     fun `딜러의 카드 총합이 카드를 뽑을 수 있는 총합의 최댓값 이하일 때, 카드를 더 받을 수 있다`() {
         val dealer =
             Dealer(
-                players = createMultiplePlayers(2),
                 gameInfo = createHitDealerGameInfo(),
             )
 
@@ -67,7 +42,6 @@ class DealerTest {
     fun `딜러의 카드 총합이 카드를 뽑을 수 있는 총합의 최댓값을 초과했을 때, 카드를 받지 않는다`() {
         val dealer =
             Dealer(
-                players = createMultiplePlayers(2),
                 gameInfo = createStandDealerGameInfo(),
             )
 
@@ -89,7 +63,6 @@ class DealerTest {
     fun `딜러의 카드 총합이 카드를 뽑을 수 있는 총합의 최댓값 이하인 상태에서 1을 뽑았을 때, 올바른 총합을 계산한다`() {
         val dealer =
             Dealer(
-                players = createMultiplePlayers(2),
                 gameInfo = createHitDealerGameInfo(),
             )
 
@@ -106,7 +79,6 @@ class DealerTest {
     fun `딜러의 카드 총합이 카드를 뽑을 수 있는 총합의 최댓값을 초과한 상태에서 1을 뽑았을 때, 올바른 총합을 계산한다`() {
         val dealer =
             Dealer(
-                players = createMultiplePlayers(2),
                 gameInfo = createStandDealerGameInfo(),
             )
 
@@ -124,7 +96,6 @@ class DealerTest {
     fun `숫자가 1인 카드를 뽑았을 때, 딜러의 카드 총합에 따라서 올바른 총합을 계산한다`(providedCard: Map<Card, Int>) {
         val dealer =
             Dealer(
-                players = createMultiplePlayers(2),
                 gameInfo = GameInfo("딜러", providedCard.keys),
             )
         dealer.drawCard {
